Move CLI fee payer arg into clap-utils
