+++ stake-accounts/src/arg_parser.rs
@@ -13,13 +13,7 @@ use std::ffi::OsString;
 use std::process::exit;
 
 fn fee_payer_arg<'a, 'b>() -> Arg<'a, 'b> {
-    Arg::with_name("fee_payer")
-        .long("fee-payer")
-        .required(true)
-        .takes_value(true)
-        .value_name("KEYPAIR")
-        .validator(is_valid_signer)
-        .help("Fee payer")
+    solana_clap_utils::fee_payer::fee_payer_arg().required(true)
 }
 
 fn funding_keypair_arg<'a, 'b>() -> Arg<'a, 'b> {
